+++ src/hooks/useDeviceType.tsx
@@ -1,9 +1,25 @@
-import React, { useContext } from 'react';
+import React, { useContext, useEffect, useState } from 'react';
+import { getDeviceType } from 'src/utils/common';
 
-export const DeviceTypeContext = React.createContext({
+interface Context {
+  deviceType: string;
+  isMobile: boolean;
+}
+
+export const DeviceTypeContext = React.createContext<Context>({
   deviceType: 'mobile',
   isMobile: true,
 });
-export const useDeviceType = () => useContext(DeviceTypeContext);
 
-export default DeviceTypeContext;
+export const DeviceTypeProvider: React.FC = (props) => {
+  const [deviceType, setDeviceType] = useState('mobile');
+  useEffect(() => {
+    const device = ['mobile', 'tablet'].includes(getDeviceType()) ? 'mobile' : 'pc';
+    setDeviceType(device);
+  }, []);
+  return (
+    <DeviceTypeContext.Provider value={{ deviceType, isMobile: deviceType === 'mobile' }}>{props.children}</DeviceTypeContext.Provider>
+  );
+};
+
+export const useDeviceType = () => useContext(DeviceTypeContext);
