+++ src/pages/[genre]/index.tsx
@@ -22,7 +22,7 @@ import { useEventTracker } from 'src/hooks/useEventTracker';
 import { RootState } from 'src/store/config';
 import { css } from '@emotion/core';
 
-import { DeviceType } from 'src/components/Context/DeviceType';
+import { DeviceTypeProvider } from 'src/hooks/useDeviceType';
 
 const { captureException } = sentry();
 
@@ -130,7 +130,7 @@ export const Home: NextPage<HomeProps> = (props) => {
         <title>{`${titleGenerator(genre)} - 리디북스`}</title>
       </Head>
       <GenreTab currentGenre={genre} />
-      <DeviceType>
+      <DeviceTypeProvider>
         {branches && branches.map((section, index) => (
           <React.Fragment key={index}>
             <HomeSectionRenderer section={section} order={index} genre={genre} />
@@ -141,7 +141,7 @@ export const Home: NextPage<HomeProps> = (props) => {
             margin-bottom: 24px;
           `}
         />
-      </DeviceType>
+      </DeviceTypeProvider>
     </>
   );
 };
